requirements.txt
amqp==5.1.1
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
blockfrost-python==0.5.3
cbor2==5.4.6
celery==5.2.7
certifi==2022.12.7
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cose==0.9.dev8
cryptography==40.0.2
Django==4.2
djangorestframework==3.14.0
ecdsa==0.18.0
ECPy==1.2.5
idna==3.4
kombu==5.2.4
mnemonic==0.20
oscrypto==1.3.0
pprintpp==0.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycardano==0.8.1
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3
redis==4.5.4
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
typeguard==2.13.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.5.1
